Refactor car and armor tracking system for improved accuracy and performance
- Added methods to draw cars and calculate their centers in the Car class.
- Enhanced the updateFromTrackedArmors method to better track armor states.
- Changed data types from float to double for improved precision across various classes including Car, Filter, Motion, Recognition, Shooter, Tracker, Trajectory, and Transformation.
- Updated the Kalman filter initialization to use double precision matrices.
- Modified the trajectory calculation to handle double precision for bullet speed and gravity.
- Improved the solvePnP and projection methods to work with double precision points.
- Adjusted the main loop to call the new drawing and center calculation methods for cars.

+++ inc/motion.hpp
@@ -7,12 +7,12 @@
 
 namespace Motion
 {
-    float yaw = 0.0f; //yaw角，单位度
-    float pitch = 0.0f; //pitch角，单位度
-    float roll = 0.0f; //roll角，单位度
-    float w_yaw_max = 180.0f; //yaw最大角速度，单位度/s
-    float w_pitch_max = 180.0f; //pitch最大角速度，单位度/s
-    cv::Point3f direction;
+    double yaw = 0.0f; //yaw角，单位度
+    double pitch = 0.0f; //pitch角，单位度
+    double roll = 0.0f; //roll角，单位度
+    double w_yaw_max = 180.0f; //yaw最大角速度，单位度/s
+    double w_pitch_max = 180.0f; //pitch最大角速度，单位度/s
+    cv::Point3d direction;
 }
