+++ inc/trajectory.hpp
@@ -5,33 +5,33 @@
 
 namespace
 {
-    float bulletSpeed = 25.0f * 1000.0f; //mm/s
-    float gravity = 9.81f * 1000.0f; //mm/s^2
+    double bulletSpeed = 25.0f * 1000.0f; //mm/s
+    double gravity = 9.81f * 1000.0f; //mm/s^2
 }
 
 
 namespace Trajectory
 {
     //规定vec3f的分量为yaw,pitch,总路径消耗时间
-    cv::Vec3f calculateTvec(cv::Point3f target_position)
+    cv::Vec3f calculateTvec(cv::Point3d target_position)
     {
         //考虑PnP解算系x轴向右，y轴向下，z轴向里
-        float yaw = -std::atan2(target_position.x, target_position.z) * 180.0f / CV_PI; //x < 0 ,z > 0时，yaw > 0
-        float distance = std::sqrt(target_position.x * target_position.x + target_position.z * target_position.z);
+        double yaw = -std::atan2(target_position.x, target_position.z) * 180.0f / CV_PI; //x < 0 ,z > 0时，yaw > 0
+        double distance = std::sqrt(target_position.x * target_position.x + target_position.z * target_position.z);
         
         // 检查抛物线是否有解：当 gravity * distance > bulletSpeed^2 时无解
-        float discriminant = gravity * distance / (bulletSpeed * bulletSpeed);
+        double discriminant = gravity * distance / (bulletSpeed * bulletSpeed);
         if(discriminant > 1.0f)
         {
             // 抛物线无解，目标距离太远
             return cv::Vec3f(0.0f, 0.0f, 0.0f);
         }
         
-        float pitch1 = atan2(gravity * distance, bulletSpeed * bulletSpeed * 2.0f) * 180.0f / CV_PI; //抛物线最高点在目标前方时，pitch > 0
-        float pitch2 = 90.0f - pitch1; //抛物线最高点在目标后方时，pitch < 0
+        double pitch1 = atan2(gravity * distance, bulletSpeed * bulletSpeed * 2.0f) * 180.0f / CV_PI; //抛物线最高点在目标前方时，pitch > 0
+        double pitch2 = 90.0f - pitch1; //抛物线最高点在目标后方时，pitch < 0
         
-        float t1 = distance / (bulletSpeed * std::cos(pitch1 * CV_PI / 180.0f));
-        float t2 = distance / (bulletSpeed * std::cos(pitch2 * CV_PI / 180.0f));
+        double t1 = distance / (bulletSpeed * std::cos(pitch1 * CV_PI / 180.0f));
+        double t2 = distance / (bulletSpeed * std::cos(pitch2 * CV_PI / 180.0f));
         
         // 检查时间是否为正值
         if(t1 <= 0.0f && t2 <= 0.0f)
